requirements.txt
Django>=3.2.0,<4.2.0
djangorestframework>=3.12.0,<3.15.0
